smart_deer_deterrent/requirements.txt
# Core dependencies
Flask==2.3.2
opencv-python==4.8.0.74
ultralytics==8.0.200
numpy==1.24.3
Pillow==10.0.0

# System monitoring
psutil==5.9.5

# Video processing
imageio==2.31.1
imageio-ffmpeg==0.4.8

# Utilities
werkzeug==2.3.6
python-dateutil==2.8.2
qrcode==7.4.2

# Optional for production
gunicorn==21.2.0

# Azure storage
azure-storage-blob==12.19.0
python-dotenv==1.0.0